refactor: 전략 주입 방식을 생성자 주입에서 setter 를 활용하는 방식으로 개선
