+++ src/main/java/lotto/domain/strategy/WinningLottoGenerateStrategy.java
@@ -1,6 +1,8 @@
-package lotto.domain;
+package lotto.domain.strategy;
 
 import camp.nextstep.edu.missionutils.Console;
+import lotto.domain.Lotto;
+import lotto.domain.strategy.LottoGenerateStrategy;
 import lotto.util.LottoParser;
 
 public class WinningLottoGenerateStrategy implements LottoGenerateStrategy {
